+++ src/main/java/transfarmer/soulweapons/item/ItemSoulDagger.java
@@ -1,6 +1,8 @@
 package transfarmer.soulweapons.item;
 
+import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.EnumAction;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.EnumActionResult;
@@ -18,23 +20,38 @@ public class ItemSoulDagger extends ItemSoulWeapon {
     }
 
     @Override
-    public ActionResult<ItemStack> onItemRightClick(World world, EntityPlayer player, EnumHand hand) {
-        if (!world.isRemote) {
-            if (player.getCapability(CAPABILITY, null).getDatum(SKILLS, DAGGER) >= 1) {
-                final ItemStack itemStack = player.getHeldItem(hand);
-                final EntitySoulDagger dagger = new EntitySoulDagger(world, player, itemStack);
+    public int getMaxItemUseDuration(final ItemStack itemStack) {
+        return 1200;
+    }
 
-                if (!player.isCreative()) {
-                    player.inventory.deleteStack(itemStack);
-                }
+    @Override
+    public EnumAction getItemUseAction(final ItemStack itemStack) {
+        return EnumAction.BOW;
+    }
 
-                dagger.shoot(player, player.rotationPitch, player.rotationYaw, 0, 1.5F, 1);
-                world.spawnEntity(dagger);
-
-                return new ActionResult<>(EnumActionResult.SUCCESS, player.getHeldItem(hand));
-            }
+    @Override
+    public ActionResult<ItemStack> onItemRightClick(final World world, final EntityPlayer player, final EnumHand hand) {
+        if (!world.isRemote && player.getCapability(CAPABILITY, null).getDatum(SKILLS, DAGGER) >= 1) {
+            player.setActiveHand(hand);
+            return new ActionResult<>(EnumActionResult.SUCCESS, player.getHeldItem(hand));
         }
 
         return new ActionResult<>(EnumActionResult.FAIL, player.getHeldItem(hand));
+    }
+
+    @Override
+    public void onPlayerStoppedUsing(final ItemStack itemStack, final World world, final EntityLivingBase entity, final int timeLeft) {
+        if (!world.isRemote) {
+            final EntitySoulDagger dagger = new EntitySoulDagger(world, entity, itemStack);
+            final EntityPlayer player = (EntityPlayer) entity;
+            final int timeTaken = this.getMaxItemUseDuration(itemStack) - timeLeft;
+
+            if (entity instanceof EntityPlayer && !player.isCreative()) {
+                player.inventory.deleteStack(itemStack);
+            }
+
+            dagger.shoot(entity, entity.rotationPitch, entity.rotationYaw, Math.min(2.5F, timeTaken / 10F * 2.5F), 1);
+            world.spawnEntity(dagger);
+        }
     }
 }
